Move Shamir service into sdk (too custom for general usage)
