+++ kryptom/src/commonMain/kotlin/com/icure/kryptom/crypto/CryptoService.kt
@@ -30,6 +30,4 @@ interface CryptoService {
 	val digest: DigestService
 
 	val hmac: HmacService
-
-	val shamirService: ShamirService
 }
