+++ kryptom/src/iosMain/kotlin/com/icure/kryptom/crypto/CryptoService.kt
@@ -13,6 +13,4 @@ actual val defaultCryptoService: CryptoService = object : CryptoService {
 	override val digest: DigestService = IosDigestService
 
 	override val hmac: HmacService = IosHmacService
-
-	override val shamirService: ShamirService = ShamirService(strongRandom)
 }
