+++ kryptom/src/jsMain/kotlin/com/icure/kryptom/crypto/CryptoService.kt
@@ -13,7 +13,5 @@ actual val defaultCryptoService: CryptoService = object : CryptoService {
 	override val digest: DigestService = JsDigestService
 
 	override val hmac: HmacService = JsHmacService
-
-	override val shamirService: ShamirService = ShamirService(strongRandom)
 }
