+++ kryptom/src/jvmAndAndroidMain/kotlin/com/icure/kryptom/crypto/CryptoService.kt
@@ -17,8 +17,6 @@ actual val defaultCryptoService = object : CryptoService {
 	override val digest: DigestService = JvmDigestService
 
 	override val hmac: HmacService = JvmHmacService
-
-	override val shamirService: ShamirService = ShamirService(strongRandom)
 }
 
 expect fun initCryptoProvider()
